Extract shared GitHub PR merge classifier
